@@ -1,17 +1,18 @@
 {
-  "name": "polis-react",
+  "name": "polis-client-admin",
   "version": "0.0.1",
   "repository": "github:pol-is/polisClientAdmin",
-  "description": "Polis React",
+  "description": "Polis Admin Console",
   "engines": {
-    "node": "6.10.3"
+    "node": "8.17.0"
   },
   "scripts": {
     "clean": "rimraf dist",
     "build:webpack": "npm run clean && NODE_ENV=production webpack --config webpack.config.js",
-    "build:webpack_unminified": "npm run clean && NODE_ENV=production webpack --config webpack.config.unminified.js",
+    "build:webpack_unminified": "npm run clean && NODE_ENV=development webpack --config webpack.config.unminified.js",
     "build": "npm run build:webpack",
     "build:preprod": "gulp deployPreprod",
+    "build-debug": "npm run build:webpack_unminified",
     "deploy:prod": "./deploy_TO_PRODUCTION",
     "deploy:preprod": "./deployPreprod",
     "start": "./x",
@@ -20,80 +21,82 @@
   },
   "author": "Colin Megill",
   "devDependencies": {
-    "babel-core": "^5.8.25",
-    "babel-eslint": "^4.1.3",
-    "babel-jest": "^17.0.0",
-    "babel-loader": "^5.3.2",
+    "@babel/core": "^7.10.3",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-decorators": "^7.10.3",
+    "@babel/plugin-transform-runtime": "^7.9.6",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.10.4",
+    "@mdx-js/loader": "^1.6.10",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-lodash": "^3.3.4",
     "babel-plugin-react-transform": "^1.1.1",
-    "babel-plugin-syntax-decorators": "^6.13.0",
-    "babel-plugin-transform-class-properties": "^6.18.0",
-    "babel-plugin-transform-es2015-computed-properties": "^6.8.0",
-    "babel-plugin-transform-object-rest-spread": "^6.16.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "es6-promise": "~1.0.0",
-    "eslint": "3.11.1",
-    "eslint-config-defaults": "^7.0.1",
-    "eslint-plugin-filenames": "^0.1.2",
-    "eslint-plugin-react": "^3.5.1",
+    "babel-plugin-transform-decorators": "^6.24.1",
+    "bluebird": "^3.7.2",
+    "es6-promise": "~4.2.8",
+    "eslint": "^7.3.1",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-config-prettier-standard": "^3.0.1",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.20.3",
+    "eslint-plugin-standard": "^4.0.1",
     "express": "^4.13.3",
     "file-loader": "https://registry.npmjs.org/file-loader/-/file-loader-0.8.5.tgz",
     "glob": "^7.1.2",
     "gulp": "^3.9.0",
     "gulp-cli": "^1.4.0",
-    "gulp-gzip": "0.0.4",
-    "gulp-rename": "~1.0.0",
+    "gulp-gzip": "1.4.2",
     "gulp-s3": "git://github.com/mbjorkegren/gulp-s3#847e2b372204b4ceeccb0e9b567f28ff00e7fa1d",
     "gulp-scp2": "^0.2.0",
-    "jest": "^17.0.0",
     "json-loader": "^0.5.1",
+    "lodash-webpack-plugin": "^0.11.5",
+    "prettier": "2.0.5",
+    "prettier-config-standard": "^1.0.1",
     "react-test-renderer": "^15.3.2",
-    "react-transform-catch-errors": "1.0.0",
     "react-transform-hmr": "1.0.1",
-    "redbox-react": "1.1.1",
     "rimraf": "^2.4.3",
     "run-sequence": "~0.3.6",
     "scp2": "^0.5.0",
-    "webpack": "^1.12.2",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.4.1"
+    "webpack": "^4.43.0",
+    "webpack-bundle-analyzer": "^3.8.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-middleware": "^1.2.0"
   },
   "dependencies": {
-    "boolean": "^0.1.3",
-    "color": "^0.7.3",
-    "highlight.js": "^8.9.1",
-    "history": "^1.12.5",
+    "@babel/runtime": "^7.10.4",
+    "@material-ui/core": "^4.9.14",
+    "boolean": "^3.0.1",
+    "d3": "^5.16.0",
     "jquery": "^2.1.4",
-    "lodash": "^3.10.1",
-    "material-ui": "^0.13.4",
-    "moment": "^2.10.6",
-    "prop-types": "^15.5.10",
-    "radium": "^0.15.0",
-    "react": "^15.5.4",
-    "react-codemirror": "^0.2.1",
-    "react-dom": "^15.5.4",
-    "react-fontawesome": "^0.3.3",
-    "react-highlight": "^0.5.1",
-    "react-markdown": "^1.1.0",
-    "react-redux": "^4.0.0",
-    "react-router": "^2.0.0-rc4",
-    "react-s-alert": "^1.3.0",
-    "react-sidebar": "^2.0.0",
-    "react-stripe-elements": "0.0.2",
-    "react-tap-event-plugin": "^0.2.1",
+    "lodash": "^4.17.19",
+    "natives": "^1.1.6",
+    "prop-types": "^15.7.2",
+    "react": "^16.4.0",
+    "react-addons-pure-render-mixin": "^15.6.2",
+    "react-addons-update": "^15.6.2",
+    "react-dom": "^16.4.0",
+    "react-easy-emoji": "^1.4.0",
+    "react-fontawesome": "^1.7.1",
+    "react-markdown": "4.3.1",
+    "react-redux": "7.2.0",
+    "react-router": "5.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-tap-event-plugin": "^2.0.0",
+    "react-transition-group": "^4.4.1",
     "redux": "^3.0.2",
     "redux-actions": "^0.9.0",
-    "redux-devtools": "^2.1.5",
     "redux-logger": "^2.0.4",
-    "redux-router": "^1.0.0-beta3",
-    "redux-thunk": "^1.0.0",
+    "redux-thunk": "^2.3.0",
     "request": "^2.65.0",
-    "tachyons-react-social-icons": "^2.8.4",
-    "victory-axis": "1.5.7",
-    "victory-bar": "2.5.0",
-    "victory-chart": "2.1.1",
-    "victory-line": "0.5.6",
-    "victory-pie": "0.1.4",
+    "theme-ui": "^0.3.1",
+    "victory": "^34.3.11",
     "whatwg-fetch": "^0.10.1"
   },
   "jest": {
