@@ -1,116 +1,69 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import Features from "../../util/plan-features";
-import { lockedIcon} from "../../util/plan-features";
-import React from "react";
-import { connect } from "react-redux";
-import Radium from "radium";
-import _ from "lodash";
-import { handleZidMetadataUpdate, optimisticZidMetadataUpdateOnTyping } from "../../actions";
-import Checkbox from "material-ui/lib/checkbox";
-import ComponentHelpers from "../../util/component-helpers";
-import NoPermission from "./no-permission";
-import InputField from "material-ui/lib/text-field";
-import settings from "../../settings";
-import Spinner from "../framework/spinner";
-import Awesome from "react-fontawesome";
+/** @jsx jsx */
 
-import ModerateCommentsSeed from "./moderate-comments-seed";
-import ModerateCommentsSeedTweet from "./moderate-comments-seed-tweet";
+import React from 'react'
+import { connect } from 'react-redux'
+import {
+  handleZidMetadataUpdate,
+  optimisticZidMetadataUpdateOnTyping
+} from '../../actions'
+import ComponentHelpers from '../../util/component-helpers'
+import NoPermission from './no-permission'
+import { Heading, Box, Flex, Text, jsx } from 'theme-ui'
+import emoji from 'react-easy-emoji'
 
-
-/* check if refer came from 'new' and if it did, show modal saying 'get started by...' */
-
-const cardPadding = 10;
-const cardBorderRadius = 3;
-
-const styles = {
-  container: {
-    backgroundColor: "rgb(240,240,247)",
-    minHeight: "100vh",
-    paddingBottom: 10
-  },
-  configCard: {
-    margin: 10,
-    maxWidth: 400,
-    backgroundColor: "rgb(253,253,253)",
-    borderRadius: cardBorderRadius,
-    padding: cardPadding,
-    WebkitBoxShadow: "3px 3px 6px -1px rgba(220,220,220,1)",
-    BoxShadow: "3px 3px 6px -1px rgba(220,220,220,1)"
-  },
-  sectionHeader: {
-    fontSize: 22,
-    marginTop: 0,
-    marginBottom: 0,
-    fontWeight: 500,
-    color: "rgb(160,160,160)"
-  },
-  notification: {
-    fontSize: 16,
-    fontWeight: 500,
-    color: "rgb(160,160,160)"
-  }
-}
+import ModerateCommentsSeed from './seed-comment'
+// import ModerateCommentsSeedTweet from "./seed-tweet";
 
 @connect(state => state.user)
 @connect(state => state.zid_metadata)
-@Radium
 class ConversationConfig extends React.Component {
-
-  handleBoolValueChange (field) {
+  handleBoolValueChange(field) {
     return () => {
-      var val = this.refs[field].isChecked();
-      if (field === "bgcolor") {
+      var val = this[field].checked
+      if (field === 'bgcolor') {
         // gray checked=default, unchecked white
-        val = val ? "default" : "#fff";
+        val = val ? 'default' : '#fff'
       }
       this.props.dispatch(
-        handleZidMetadataUpdate(
-          this.props.zid_metadata,
-          field,
-          val
-        )
+        handleZidMetadataUpdate(this.props.zid_metadata, field, val)
       )
     }
   }
 
   transformBoolToInt(value) {
-    return value ? 1 : 0;
+    return value ? 1 : 0
   }
 
-  handleIntegerBoolValueChange (field) {
+  handleIntegerBoolValueChange(field) {
     return () => {
       this.props.dispatch(
         handleZidMetadataUpdate(
           this.props.zid_metadata,
           field,
-          this.transformBoolToInt(this.refs[field].isChecked())
+          this.transformBoolToInt(this[field].checked)
         )
       )
     }
   }
 
-  handleStringValueChange (field) {
+  handleStringValueChange(field) {
     return () => {
-      var val = this.refs[field].getValue();
-      if (field === "help_bgcolor" || field === "help_color") {
+      let val = this[field].value
+      if (field === 'help_bgcolor' || field === 'help_color') {
         if (!val.length) {
-          val = "default";
+          val = 'default'
         }
       }
       this.props.dispatch(
-        handleZidMetadataUpdate(
-          this.props.zid_metadata,
-          field,
-          this.refs[field].getValue()
-        )
+        handleZidMetadataUpdate(this.props.zid_metadata, field, val)
       )
     }
   }
 
-  handleConfigInputTyping (field) {
-    return (e) => {
+  handleConfigInputTyping(field) {
+    return e => {
       this.props.dispatch(
         optimisticZidMetadataUpdateOnTyping(
           this.props.zid_metadata,
@@ -122,243 +75,319 @@ class ConversationConfig extends React.Component {
   }
 
   render() {
-    console.log(this.props);
-
     if (ComponentHelpers.shouldShowPermissionsError(this.props)) {
-      return <NoPermission/>
+      return <NoPermission />
     }
 
-    const canToggleVis = Features.canToggleVisVisibility(this.props.user);
-    const canToggleCommentForm = Features.canToggleCommentForm(this.props.user);
-    const canToggleStrictMod = Features.canToggleStrictMod(this.props.user);
-    const canCustomizeColors = Features.canCustomizeColors(this.props.user);
-
     return (
-      <div style={styles.container}>
-        <div style={styles.configCard}>
-        {
-          this.props.loading ?
-            <div style={styles.notification}> <Awesome name="cloud-upload"/> <span>Saving</span> </div> :
-            <div style={styles.notification}> <Awesome name="bolt"/> <span>Up to date</span> </div>
-        }
-        {this.props.error ? <div style={styles.notification}> <Awesome name="exclamation-circle"/> Error Saving </div> : ""}
-        </div>
-        <div style={styles.configCard}>
-          <p style={styles.sectionHeader}>Topic</p>
-          <InputField
-            ref={"topic"}
-            style={{width: 360}}
-            onBlur={this.handleStringValueChange("topic").bind(this)}
-            floatingLabelText="Topic"
-            onChange={this.handleConfigInputTyping("topic")}
-            value={this.props.zid_metadata.topic}
-            multiLine={true} />
-        </div>
-        <div style={styles.configCard}>
-          <p style={styles.sectionHeader}>Description</p>
-          <InputField
-            hintText="Can include markdown!"
-            style={{width: 360}}
-            ref={"description"}
-            onBlur={this.handleStringValueChange("description").bind(this)}
-            floatingLabelText="Description"
-            onChange={this.handleConfigInputTyping("description")}
-            value={this.props.zid_metadata.description}
-            multiLine={true} />
-        </div>
+      <Box>
+        <Heading
+          as="h3"
+          sx={{
+            fontSize: [3, null, 4],
+            lineHeight: 'body',
+            mb: [3, null, 4]
+          }}>
+          Configure
+        </Heading>
+        <Box sx={{ mb: [4] }}>
+          {this.props.loading ? (
+            <Text>{emoji('💾')} Saving</Text>
+          ) : (
+            <Text>{emoji('⚡')} Up to date</Text>
+          )}
+          {this.props.error ? <Text>Error Saving</Text> : null}
+        </Box>
+        <Box sx={{ mb: [3] }}>
+          <Text sx={{ mb: [2] }}>Topic</Text>
+          <input
+            ref={c => (this.topic = c)}
+            sx={{
+              fontFamily: 'body',
+              fontSize: [2],
+              width: '35em',
+              borderRadius: 2,
+              padding: [2],
+              border: '1px solid',
+              borderColor: 'mediumGray'
+            }}
+            onBlur={this.handleStringValueChange('topic').bind(this)}
+            onChange={this.handleConfigInputTyping('topic').bind(this)}
+            defaultValue={this.props.zid_metadata.topic}
+          />
+        </Box>
 
-        <ModerateCommentsSeed params={{conversation_id: this.props.zid_metadata.conversation_id}}/>
-        <ModerateCommentsSeedTweet params={{conversation_id: this.props.zid_metadata.conversation_id}}/>
+        <Box sx={{ mb: [3] }}>
+          <Text sx={{ mb: [2] }}>Description</Text>
+          <textarea
+            ref={c => (this.description = c)}
+            sx={{
+              fontFamily: 'body',
+              fontSize: [2],
+              width: '35em',
+              height: '7em',
+              resize: 'none',
+              padding: [2],
+              borderRadius: 2,
+              border: '1px solid',
+              borderColor: 'mediumGray'
+            }}
+            onBlur={this.handleStringValueChange('description').bind(this)}
+            onChange={this.handleConfigInputTyping('description').bind(this)}
+            defaultValue={this.props.zid_metadata.description}
+          />
+        </Box>
 
-        <div style={styles.configCard}>
-          <p style={styles.sectionHeader}> Customize the User Interface </p>
-          <div style={{marginTop: 20}}> </div>
-          <Checkbox
-            label= {"Visualization" + (canToggleVis ? "" : lockedIcon)}
-            disabled={!canToggleVis}
-            ref={"vis_type"}
-            checked={ this.props.zid_metadata.vis_type === 1 ? true : false }
-            onCheck={ this.handleIntegerBoolValueChange("vis_type").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> participants can see the visualization </p>
-          <Checkbox
-            label={"Comment form" + (canToggleCommentForm ? "" : lockedIcon)}
-            ref={"write_type"}
-            disabled={!canToggleCommentForm}
-            checked={this.props.zid_metadata.write_type === 1 ? true : false}
-            onCheck={ this.handleIntegerBoolValueChange("write_type").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> Users can submit comments </p>
-          {false ? (<span><Checkbox
-            label="Voting pane"
-            ref={"upvotes"}
-            disabled
-            checked={this.props.zid_metadata.upvotes === 1 ? true : false}
-            onCheck={ this.handleIntegerBoolValueChange("upvotes").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> Users can vote on comments </p></span>) : null}
-          <Checkbox
-            label="Help text"
-            ref={"help_type"}
-            checked={this.props.zid_metadata.help_type === 1 ? true : false}
-            onCheck={ this.handleIntegerBoolValueChange("help_type").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> Show the two explanation modals above voting and the visualization </p>
-          <Checkbox
-            label="Prompt participants to subscribe to updates"
-            ref={"subscribe_type"}
-            checked={this.props.zid_metadata.subscribe_type === 1 ? true : false}
-            onCheck={ this.handleIntegerBoolValueChange("subscribe_type").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> A prompt is shown to users once they finish voting on all available comments. They may optionally provide their email address to receive notifications when there are new comments to vote on. </p>
-          <Checkbox
-            label="Social sharing buttons"
-            ref={"socialbtn_type"}
-            checked={this.props.zid_metadata.socialbtn_type === 1 ? true : false}
-            onCheck={ this.handleIntegerBoolValueChange("socialbtn_type").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> </p>
-          <Checkbox
-            label="Facebook login prompt"
-            ref={"auth_opt_fb"}
-            checked={this.props.zid_metadata.auth_opt_fb}
-            onCheck={ this.handleBoolValueChange("auth_opt_fb").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> </p>
-          <Checkbox
-            label="Twitter login prompt"
-            ref={"auth_opt_tw"}
-            checked={this.props.zid_metadata.auth_opt_tw}
-            onCheck={ this.handleBoolValueChange("auth_opt_tw").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> </p>
-          {false ? (<span><Checkbox
-            label="Gray background"
-            ref={"bgcolor"}
-            checked={this.props.zid_metadata.bgcolor === null ? true : false}
-            onCheck={ this.handleBoolValueChange("bgcolor").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}>
-              {"Unchecked: white background"}
-            </p></span>) : null
-          }
-          <div>
-            <InputField
-              ref={"style_btn"}
-              disabled={!canCustomizeColors}
-              style={{width: 360}}
-              onBlur={this.handleStringValueChange("style_btn").bind(this)}
-              hintText="ie., #e63082"
-              onChange={this.handleConfigInputTyping("style_btn")}
-              value={this.props.zid_metadata.style_btn}
-              floatingLabelText={"Customize submit button color" + (canCustomizeColors ? "" : lockedIcon)}
-              multiLine={true} />
-          </div>
-          <div>
-            <InputField
-              ref={"help_bgcolor"}
-              disabled={!canCustomizeColors}
-              style={{width: 360}}
-              onBlur={this.handleStringValueChange("help_bgcolor").bind(this)}
-              onChange={this.handleConfigInputTyping("help_bgcolor")}
-              value={this.props.zid_metadata.help_bgcolor}
-              hintText="ie., #e63082"
-              floatingLabelText={"Customize help text background" + (canCustomizeColors ? "" : lockedIcon)}
-              multiLine={true} />
-          </div>
-          <div>
-            <InputField
-              ref={"help_color"}
-              disabled={!canCustomizeColors}
-              style={{width: 360}}
-              onBlur={this.handleStringValueChange("help_color").bind(this)}
-              onChange={this.handleConfigInputTyping("help_color")}
-              value={this.props.zid_metadata.help_color}
-              hintText="ie., #e63082"
-              floatingLabelText={"Customize help text color" + (canCustomizeColors ? "" : lockedIcon)}
-              multiLine={true} />
-          </div>
-        </div>
-        <div style={styles.configCard}>
-          <p style={styles.sectionHeader}> Schemes </p>
-          <div style={{marginTop: 20}}> </div>
-          <Checkbox
-            label={"Strict Moderation" + (canToggleStrictMod ? "" : lockedIcon)}
-            ref={"strict_moderation"}
-            disabled={!canToggleStrictMod}
-            checked={ this.props.zid_metadata.strict_moderation }
-            onCheck={ this.handleBoolValueChange("strict_moderation").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> no comments shown without moderator approval </p>
-          <Checkbox
-            label="Require Auth to Comment"
-            ref={"auth_needed_to_write"}
-            checked={this.props.zid_metadata.auth_needed_to_write}
-            onCheck={ this.handleBoolValueChange("auth_needed_to_write").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> Users cannot submit comments without first connecting either Facebook or Twitter </p>
-          <Checkbox
-            label="Require Auth to Vote"
-            ref={"auth_needed_to_vote"}
-            checked={this.props.zid_metadata.auth_needed_to_vote}
-            onCheck={ this.handleBoolValueChange("auth_needed_to_vote").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> Users cannot vote without first connecting either Facebook or Twitter </p>
-          <Checkbox
-            label="Open Data"
-            ref={"is_data_open"}
-            checked={this.props.zid_metadata.is_data_open}
-            onCheck={ this.handleBoolValueChange("is_data_open").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}
-            color={settings.polisBlue}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> Comments, votes, and group data can be exported by any user </p>
-          {false ? (<span><Checkbox
-            label="Preserve Anonymity"
-            ref={"is_anon"}
-            disabled
-            checked={this.props.zid_metadata.is_anon}
-            onCheck={ this.handleBoolValueChange("is_anon").bind(this) }
-            labelPosition={"left"}
-            labelWrapperColor={settings.darkerGray}/>
-            <p style={{fontSize: 10, fontStyle: "italic"}}> Disables visualization, does not transmit any participant statistics to you, requires social authorization for both writing and voting. </p></span>) : ""}
-        </div>
-      </div>
-    );
+        <Heading
+          as="h6"
+          sx={{
+            fontSize: [1, null, 2],
+            lineHeight: 'body',
+            my: [3, null, 4]
+          }}>
+          Seed Comments
+        </Heading>
+        <ModerateCommentsSeed
+          params={{ conversation_id: this.props.zid_metadata.conversation_id }}
+        />
+
+        {/* <ModerateCommentsSeedTweet
+          params={{ conversation_id: this.props.zid_metadata.conversation_id }}
+        /> */}
+
+        <Heading
+          as="h6"
+          sx={{
+            fontSize: [1, null, 2],
+            lineHeight: 'body',
+            my: [3, null, 4]
+          }}>
+          Customize the user interface
+        </Heading>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Visualization"
+              ref={c => (this.vis_type = c)}
+              checked={this.props.zid_metadata.vis_type === 1}
+              onChange={this.handleIntegerBoolValueChange('vis_type').bind(
+                this
+              )}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>Participants can see the visualization</Text>
+          </Box>
+        </Flex>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Comment form"
+              ref={c => (this.write_type = c)}
+              checked={this.props.zid_metadata.write_type === 1}
+              onChange={this.handleIntegerBoolValueChange('write_type').bind(
+                this
+              )}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>Participants can submit comments</Text>
+          </Box>
+        </Flex>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Help text"
+              ref={c => (this.help_type = c)}
+              checked={this.props.zid_metadata.help_type === 1}
+              onChange={this.handleIntegerBoolValueChange('help_type').bind(
+                this
+              )}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>Show explanation text above voting and visualization</Text>
+          </Box>
+        </Flex>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Prompt participants to subscribe to updates"
+              ref={c => (this.subscribe_type = c)}
+              checked={this.props.zid_metadata.subscribe_type === 1}
+              onChange={this.handleIntegerBoolValueChange(
+                'subscribe_type'
+              ).bind(this)}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>
+              Prompt participants to subscribe to updates. A prompt is shown to
+              users once they finish voting on all available comments. If
+              enabled, participants may optionally provide their email address
+              to receive notifications when there are new comments to vote on.
+            </Text>
+          </Box>
+        </Flex>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Facebook login prompt"
+              ref={c => (this.auth_opt_fb = c)}
+              checked={this.props.zid_metadata.auth_opt_fb}
+              onChange={this.handleBoolValueChange('auth_opt_fb').bind(this)}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>Show Facebook login prompt</Text>
+          </Box>
+        </Flex>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Twitter login prompt"
+              ref={c => (this.auth_opt_tw = c)}
+              checked={this.props.zid_metadata.auth_opt_tw}
+              onChange={this.handleBoolValueChange('auth_opt_tw').bind(this)}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>Show Twitter login prompt</Text>
+          </Box>
+        </Flex>
+
+        <Heading
+          as="h6"
+          sx={{
+            fontSize: [1, null, 2],
+            lineHeight: 'body',
+            my: [3, null, 4]
+          }}>
+          Schemes
+        </Heading>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              ref={c => (this.strict_moderation = c)}
+              checked={this.props.zid_metadata.strict_moderation}
+              onChange={this.handleBoolValueChange('strict_moderation').bind(
+                this
+              )}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>No comments shown without moderator approval</Text>
+          </Box>
+        </Flex>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Require Auth to Comment"
+              ref={c => (this.auth_needed_to_write = c)}
+              checked={this.props.zid_metadata.auth_needed_to_write}
+              onChange={this.handleBoolValueChange('auth_needed_to_write').bind(
+                this
+              )}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>
+              Participants cannot submit comments without first connecting
+              either Facebook or Twitter
+            </Text>
+          </Box>
+        </Flex>
+
+        <Flex sx={{ alignItems: 'flex-start', mb: [3] }}>
+          <Box sx={{ flexShrink: 0, position: 'relative', top: -0.5 }}>
+            <input
+              type="checkbox"
+              label="Require Auth to Vote"
+              ref={c => (this.auth_needed_to_vote = c)}
+              checked={this.props.zid_metadata.auth_needed_to_vote}
+              onChange={this.handleBoolValueChange('auth_needed_to_vote').bind(
+                this
+              )}
+            />
+          </Box>
+          <Box sx={{ ml: [2], flexShrink: 0, maxWidth: '35em' }}>
+            <Text>
+              Participants cannot vote without first connecting either Facebook
+              or Twitter
+            </Text>
+          </Box>
+        </Flex>
+      </Box>
+    )
   }
 }
 
-export default ConversationConfig;
+export default ConversationConfig
 
-// <Checkbox
-//   label="Read Only"
-//   disabled
-//   checked={false}
-//   onCheck={ () => {console.log("this should be an action")} }
-//   labelPosition={"left"}
-//   labelWrapperColor={settings.darkerGray}/>
-//   <p style={{fontSize: 10, fontStyle: "italic"}}> Disables writing and commenting, enables visualization. </p>
+// checked={this.props.zid_metadata.is_data_open}
+// Comments, votes, and group data can be exported by any user
+
+/* <InputField
+            ref={"style_btn"}
+
+            style={{ width: 360 }}
+            onBlur={this.handleStringValueChange("style_btn").bind(this)}
+            hintText="ie., #e63082"
+            onChange={this.handleConfigInputTyping("style_btn")}
+            value={this.props.zid_metadata.style_btn}
+            floatingLabelText={
+              "Customize submit button color" + (canCustomizeColors ? "" : lockedIcon)
+            }
+            multiLine={true}
+          /> */
+
+/* <InputField
+            ref={"help_bgcolor"}
+
+            style={{ width: 360 }}
+            onBlur={this.handleStringValueChange("help_bgcolor").bind(this)}
+            onChange={this.handleConfigInputTyping("help_bgcolor")}
+            value={this.props.zid_metadata.help_bgcolor}
+            hintText="ie., #e63082"
+            floatingLabelText={
+              "Customize help text background" + (canCustomizeColors ? "" : lockedIcon)
+            }
+            multiLine={true}
+          /> */
+
+/* <InputField
+            ref={"help_color"}
+
+            style={{ width: 360 }}
+            onBlur={this.handleStringValueChange("help_color").bind(this)}
+            onChange={this.handleConfigInputTyping("help_color")}
+            value={this.props.zid_metadata.help_color}
+            hintText="ie., #e63082"
+            floatingLabelText={"Customize help text color" + (canCustomizeColors ? "" : lockedIcon)}
+            multiLine={true}
+          /> */
+
+/* <Checkbox
+            label="Social sharing buttons"
+            ref={"socialbtn_type"}
+            checked={this.props.zid_metadata.socialbtn_type === 1 ? true : false}
+            onCheck={this.handleIntegerBoolValueChange("socialbtn_type").bind(this)}
+
+
+          /> */
