@@ -1,197 +1,145 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import React from "react";
-import { connect } from "react-redux";
-import Radium from "radium";
-import { startDataExport } from "../../actions";
-import _ from "lodash";
-import RadioButton from "material-ui/lib/radio-button";
-import RadioButtonGroup from "material-ui/lib/radio-button-group";
-import SelectField from "material-ui/lib/select-field";
-import Button from "../framework/generic-button";
-import dateSetupUtil from "../../util/data-export-date-setup";
-import Flex from "../framework/flex";
-import Awesome from "react-fontawesome";
-import Checkbox from "material-ui/lib/checkbox";
-import settings from "../../settings";
-
-const styles = {
-  container: {
-    backgroundColor: "rgb(240,240,247)",
-    marginBottom: -20,
-  },
-  exportCard: {
-    backgroundColor: "rgb(253,253,253)",
-    margin: "0px 20px 20px 20px",
-    borderRadius: 3,
-    padding: 20,
-    WebkitBoxShadow: "3px 3px 6px -1px rgba(220,220,220,1)",
-    BoxShadow: "3px 3px 6px -1px rgba(220,220,220,1)"
-  },
-};
+import React from 'react'
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
+import { startDataExport } from '../../actions'
+import dateSetupUtil from '../../util/data-export-date-setup'
+import { Heading } from 'theme-ui'
+import ComponentHelpers from '../../util/component-helpers'
+import NoPermission from './no-permission'
 
 @connect(state => state.zid_metadata)
-@Radium
 class DataExport extends React.Component {
   constructor(props) {
-    super(props);
-    var times = dateSetupUtil();
-    this.state = Object.assign({},times,{showHelpMessage: false});
+    super(props)
+    var times = dateSetupUtil()
+    this.state = Object.assign({}, times, { showHelpMessage: false })
   }
+
   handleExportClicked() {
     return () => {
+      this.setState({ showHelpMessage: true })
 
-      this.setState({showHelpMessage: true})
+      const year = this.exportSelectYear.value
+      const month = this.exportSelectMonth.value
+      const dayOfMonth = this.exportSelectDay.value
+      const tz = this.exportSelectHour.value
+      const format = 'csv'
 
-      const year = this.refs.exportSelectYear.value;
-      const month = this.refs.exportSelectMonth.value;
-      const dayOfMonth = this.refs.exportSelectDay.value;
-      const tz = this.refs.exportSelectHour.value;
-      const format = this.refs.chooseFormat.getSelectedValue();
-      console.log("format", format);
-
-      const dateString = [year, month, dayOfMonth, tz].join(" ");
-      const dddate = new Date(dateString);
+      const dateString = [year, month, dayOfMonth, tz].join(' ')
+      const dddate = new Date(dateString)
 
       this.props.dispatch(
         startDataExport(
           this.props.zid_metadata.conversation_id,
           format,
-          ((dddate/1000) << 0),
+          (dddate / 1000) << 0,
           !!this.state.untilEnabled
         )
       )
     }
   }
+
   showHelpMessage() {
     return (
-      <div style={{marginTop: 20}}>
-        <p style={{
-            backgroundColor: "rgb(240,240,240)",
-            padding: 20,
-            borderRadius: 3
-          }}>
-          <Awesome name="cloud-download"/>
-          {" "}
-          Data from this conversation will be sent to your email.
-          (This can take a little while, especially for larger conversations).
+      <div>
+        <p>
+          Data from this conversation will be sent to your email. (This can take
+          a little while, especially for larger conversations).
         </p>
       </div>
     )
   }
+
   handleUntilToggled() {
-    this.setState({untilEnabled: !this.state.untilEnabled});
+    this.setState({ untilEnabled: !this.state.untilEnabled })
   }
+
   render() {
+    if (ComponentHelpers.shouldShowPermissionsError(this.props)) {
+      return <NoPermission />
+    }
+
     return (
-      <div style={styles.container}>
-        <div style={styles.exportCard}>
-          <p style={{marginTop: 0}}> Until: </p>
-          <input
-            onClick={this.handleUntilToggled.bind(this)}
-            style={{
-              marginRight: 10,
-            }}
-            type="checkbox"/>
+      <div>
+        <Heading
+          as="h3"
+          sx={{
+            fontSize: [3, null, 4],
+            lineHeight: 'body',
+            mb: [3, null, 4]
+          }}>
+          Export
+        </Heading>
+        <div>
+          <p> Until: </p>
+          <input onClick={this.handleUntilToggled.bind(this)} type="checkbox" />
           <select
-            disabled={this.state.untilEnabled ? "" : "disabled"}
-            style={{
-              marginRight: 10,
-              cursor: "pointer",
-              fontSize: 16,
-            }}
-            ref="exportSelectYear">
-            {
-              this.state.years.map((year, i) => {
-                return (
-                  <option selected={year.selected} key={i} value={year.name}> {year.name} </option>
-                )
-              })
-            }
+            disabled={this.state.untilEnabled ? '' : 'disabled'}
+            ref={c => (this.exportSelectYear = c)}>
+            {this.state.years.map((year, i) => {
+              return (
+                <option selected={year.selected} key={i} value={year.name}>
+                  {year.name}
+                </option>
+              )
+            })}
           </select>
           <select
-            disabled={this.state.untilEnabled ? "" : "disabled"}
-            style={{
-              marginRight: 10,
-              cursor: "pointer",
-              fontSize: 16,
-
-            }}
-            ref="exportSelectMonth">
-            {
-              this.state.months.map((month, i)=>{
-                return (
-                  <option selected={month.selected} key={i} value={month.name}> {month.name} </option>
-                )
-              })
-            }
+            disabled={this.state.untilEnabled ? '' : 'disabled'}
+            ref={c => (this.exportSelectMonth = c)}>
+            {this.state.months.map((month, i) => {
+              return (
+                <option selected={month.selected} key={i} value={month.name}>
+                  {month.name}
+                </option>
+              )
+            })}
           </select>
           <select
-            disabled={this.state.untilEnabled ? "" : "disabled"}
-            style={{
-              marginRight: 10,
-              cursor: "pointer",
-              fontSize: 16,
-
-            }}
-            ref="exportSelectDay">
-            {
-              this.state.days.map((day, i)=>{
-                return (
-                  <option selected={day.selected} key={i} value={day.name}> {day.name} </option>
-                );
-              })
-            }
+            disabled={this.state.untilEnabled ? '' : 'disabled'}
+            ref={c => (this.exportSelectDay = c)}>
+            {this.state.days.map((day, i) => {
+              return (
+                <option selected={day.selected} key={i} value={day.name}>
+                  {' '}
+                  {day.name}{' '}
+                </option>
+              )
+            })}
           </select>
           <select
-            disabled={this.state.untilEnabled ? "" : "disabled"}
-            style={{
-              marginRight: 10,
-              cursor: "pointer",
-              fontSize: 16,
-            }}
-            ref="exportSelectHour">
-            {
-              this.state.tzs.map((tz, i) => {
-                return (
-                  <option selected={tz.selected} key={i} value={tz.name}> {tz.name} </option>
-                );
-              })
-            }
+            disabled={this.state.untilEnabled ? '' : 'disabled'}
+            ref={c => (this.exportSelectHour = c)}>
+            {this.state.tzs.map((tz, i) => {
+              return (
+                <option selected={tz.selected} key={i} value={tz.name}>
+                  {' '}
+                  {tz.name}{' '}
+                </option>
+              )
+            })}
           </select>
-          <p style={{fontWeight: 300, maxWidth: 400, lineHeight: 1.5}}> {"(By default, the entire dataset is returned. To limit the last timestamp returned, enter a date here.)"} </p>
-          <p> Format: </p>
-          <RadioButtonGroup ref="chooseFormat" name="format" defaultSelected="csv">
-            <RadioButton
-              value="csv"
-              label="CSV"
-              style={{marginBottom:5}} />
-            <RadioButton
-              disabled
-              value="xls"
-              label="XLS (coming soon)"
-              style={{marginBottom:5}}/>
-          </RadioButtonGroup>
-          <Button
-            style={{
-              backgroundColor: "#03a9f4",
-              color: "white",
-              marginTop: 20,
-            }}
-            onClick={this.handleExportClicked().bind(this)}
-            >
+          <p>
+            By default, the entire dataset is returned. To limit the last
+            timestamp returned, enter a date here.
+          </p>
+          <button onClick={this.handleExportClicked().bind(this)}>
             Export
-          </Button>
-          {this.state.showHelpMessage ? this.showHelpMessage() : ""}
+          </button>
+          {this.state.showHelpMessage ? this.showHelpMessage() : null}
         </div>
       </div>
-    );
+    )
   }
 }
 
-export default DataExport;
+DataExport.propTypes = {
+  dispatch: PropTypes.func,
+  zid_metadata: PropTypes.shape({
+    conversation_id: PropTypes.string
+  })
+}
 
-/*
-  todo
-    validation of dates with moment - should have the start date as well
-*/
+export default DataExport
