@@ -1,11 +1,9 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import React from "react";
-import Radium from "radium";
+import React from 'react'
 
-@Radium
 class Spinner extends React.Component {
-  getSpinner () {
+  getSpinner() {
     return {
       __html: `
         <svg
@@ -223,13 +221,12 @@ class Spinner extends React.Component {
             to="0" dur="1s" begin="0.9166666666666666s" repeatCount="indefinite"/>
           </rect>
         </svg>`
-    };
+    }
   }
+
   render() {
-    return (
-      <div dangerouslySetInnerHTML={this.getSpinner()} />
-    );
+    return <div dangerouslySetInnerHTML={this.getSpinner()} />
   }
 }
 
-export default Spinner;
+export default Spinner
