@@ -1,17 +1,17 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import * as types from "../actions";
-
-const conversation_config = (state = {
-  loading: false,
-  conversation_config: null,
-  update_error: null
-}, action) => {
+const conversation_config = (
+  state = {
+    loading: false,
+    conversation_config: null,
+    update_error: null
+  },
+  action
+) => {
   switch (action.type) {
-
-  default:
-    return state;
+    default:
+      return state
   }
-};
+}
 
-export default conversation_config;
+export default conversation_config
