@@ -1,37 +1,37 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 const settings = {
-  polisBlue: "rgb(81, 171, 244)",
-  textColor: "#3e3e3e",
-  linkColor: "#777777",
-  darkColorBase: "#202020",
-  lightColorBase: "#F1F1F1",
-  clickCursor: "pointer",
-  linkHoverColor: "#3e3e3e",
-  highlightColor: "#006FC7",
-  transitionAll: "all ease .3s",
-  bgColorOffset: "#f8f8f8",
-  bgColorBase: "#ffffff",
-  borderColorBase: "#BEBEBE",
+  polisBlue: 'rgb(81, 171, 244)',
+  textColor: '#3e3e3e',
+  linkColor: '#777777',
+  darkColorBase: '#202020',
+  lightColorBase: '#F1F1F1',
+  clickCursor: 'pointer',
+  linkHoverColor: '#3e3e3e',
+  highlightColor: '#006FC7',
+  transitionAll: 'all ease .3s',
+  bgColorOffset: '#f8f8f8',
+  bgColorBase: '#ffffff',
+  borderColorBase: '#BEBEBE',
   borderRadiusBase: 4,
-  borderStyleBase: "solid",
+  borderStyleBase: 'solid',
   borderWidthBase: 1,
-  fontFamilyMonospace: "\"Courier New\", monospace",
-  fontFamilySansSerif: "\"Helvetica Neue\", Helvetica, Arial, sans-serif",
+  fontFamilyMonospace: '"Courier New", monospace',
+  fontFamilySansSerif: '"Helvetica Neue", Helvetica, Arial, sans-serif',
   fontSizeBase: 14,
   lineHeightBase: 1.428571429,
   lineHeightSmall: 1.3,
 
-  lightestGray: "#FFFFFF",
-  lighterGray: "#F6F6F6",
-  lightGray: "#F1F1F1",
-  mediumGray: "#E0E0E0",
-  gray: "#979B9F",
-  darkBlueGray: "#ACB1B8",
-  blueGray: "#BFC5CC",
-  darkGray: "#6C7277",
-  darkerGray: "#34383C",
-  darkestGray: "#1E2226",
-};
+  lightestGray: '#FFFFFF',
+  lighterGray: '#F6F6F6',
+  lightGray: '#F1F1F1',
+  mediumGray: '#E0E0E0',
+  gray: '#979B9F',
+  darkBlueGray: '#ACB1B8',
+  blueGray: '#BFC5CC',
+  darkGray: '#6C7277',
+  darkerGray: '#34383C',
+  darkestGray: '#1E2226'
+}
 
-export default settings;
+export default settings
