@@ -1,33 +1,16 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { createStore, applyMiddleware, compose } from "redux";
-import thunk from "redux-thunk";
-import { devTools } from 'redux-devtools';
-import createHistory from 'history/lib/createBrowserHistory';
+import { createStore, applyMiddleware, compose } from 'redux'
+import thunk from 'redux-thunk'
 
-import rootReducer from "../reducers";
+import rootReducer from '../reducers'
 
-const middleware = [thunk];
+const middleware = [thunk]
 
-let finalCreateStore;
-
-if (process.env.NODE_ENV === 'production') {
-  finalCreateStore = compose(
-    applyMiddleware(...middleware),
-    require('redux-devtools').devTools()
-  )(createStore);
-} else {
-  finalCreateStore = compose(
-    applyMiddleware(...middleware),
-    require('redux-devtools').devTools(),
-    require('redux-devtools').persistState(
-      window.location.href.match(/[?&]debug_session=([^&]+)\b/)
-    )
-  )(createStore);
-}
+const finalCreateStore = compose(applyMiddleware(...middleware))(createStore)
 
 const configureStore = function (initialState) {
-  return finalCreateStore(rootReducer, initialState);
-};
+  return finalCreateStore(rootReducer, initialState)
+}
 
-export default configureStore;
+export default configureStore
