@@ -1,23 +1,31 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-var s = {};
+var s = {}
 
 // login/createuser strings
-s.polis_err_param_parse_failed_email = "Por favor insira um endereço de email válido.";
-s.polis_err_reg_bad_email = "Por favor insira um endereço de email válido.";
-s.polis_err_param_parse_failed_password = "Por favor insira uma senha válida."
-s.polis_err_login_unknown_user_or_password_noresults = "Login falhou: usuário/senha inválida.";
-s.polis_err_login_unknown_user_or_password = "Login falhou: usuário/senha inválida.";
-s.polis_err_reg_user_with_that_email_exists = "Email já cadastrado, tente fazer login.";
-s.polis_err_reg_need_name = "Por favor inclua seu nome.";
+s.polis_err_param_parse_failed_email =
+  'Por favor insira um endereço de email válido.'
+s.polis_err_reg_bad_email = 'Por favor insira um endereço de email válido.'
+s.polis_err_param_parse_failed_password = 'Por favor insira uma senha válida.'
+s.polis_err_login_unknown_user_or_password_noresults =
+  'Login falhou: usuário/senha inválida.'
+s.polis_err_login_unknown_user_or_password =
+  'Login falhou: usuário/senha inválida.'
+s.polis_err_reg_user_with_that_email_exists =
+  'Email já cadastrado, tente fazer login.'
+s.polis_err_reg_need_name = 'Por favor inclua seu nome.'
 
-s.polis_err_post_comment_duplicate = "Erro ao publicar. Este comentário já existe!";
-s.waitinglist_add_success = "Você já foi incluído na lista de email! Manteremos contato.";
+s.polis_err_post_comment_duplicate =
+  'Erro ao publicar. Este comentário já existe!'
+s.waitinglist_add_success =
+  'Você já foi incluído na lista de email! Manteremos contato.'
 
-s.polis_err_fetching_tweet = "Erro buscando tuíte. Verifique se esse link é mesmo de um tuíte.";
+s.polis_err_fetching_tweet =
+  'Erro buscando tuíte. Verifique se esse link é mesmo de um tuíte.'
 
-s.share_but_no_comments_warning = "Essa discussão não tem comentários. Nós recomendamos adicionar alguns comentários antes de convidar participantes. Isso irá ajudar os participantes a começar. No menu da direita, vá para a opção 'comments' e depois 'seed'.";
-s.share_but_no_visible_comments_warning = "Essa discussão não tem comentários visíveis. Nós recomendamos acrescentar alguns comentários ou moderar comentários que estejam na fila antes de convidar os participantes, já que isso os ajudará a entender que tipo de comentários eles devem enviar.";
+s.share_but_no_comments_warning =
+  "Essa discussão não tem comentários. Nós recomendamos adicionar alguns comentários antes de convidar participantes. Isso irá ajudar os participantes a começar. No menu da direita, vá para a opção 'comments' e depois 'seed'."
+s.share_but_no_visible_comments_warning =
+  'Essa discussão não tem comentários visíveis. Nós recomendamos acrescentar alguns comentários ou moderar comentários que estejam na fila antes de convidar os participantes, já que isso os ajudará a entender que tipo de comentários eles devem enviar.'
 
-
-export default s;
+export default s
