@@ -1,40 +1,38 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-var path = require('path');
-var webpack = require('webpack');
+var path = require("path");
+var webpack = require("webpack");
+
+const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
 
 module.exports = {
-  devtool: ['eval','sourcemap'],
-  entry: [
-    'webpack-hot-middleware/client',
-    './src/index'
-  ],
+  // devtool: ["eval", "sourcemap"],
+  entry: ["./src/index"],
   output: {
-    path: path.join(__dirname, 'devel'),
-    filename: 'admin_bundle.js',
-    publicPath: '/dist/'
+    path: path.join(__dirname, "devel"),
+    filename: "admin_bundle.js",
+    publicPath: "/dist/",
   },
-  plugins: [
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.NoErrorsPlugin()
-  ],
+  // plugins: [new BundleAnalyzerPlugin()],
   resolve: {
-    extentions: [ '', '.js', '.css', '.png', '.svg' ]
+    extensions: [".js", ".css", ".png", ".svg"],
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.js$/,
-        loaders: ['babel'],
-        include: path.join(__dirname, 'src')
+        loaders: ["babel-loader"],
+        include: path.join(__dirname, "src"),
       },
-      {
-        test: /\.json$/, loader: "json-loader"
-      },
+
       {
         test: /\.(png|jpg|gif|svg)$/,
-        loader: "file-loader"
+        loader: "file-loader",
+      },
+      {
+        test: /\.mdx?$/,
+        use: ['babel-loader', '@mdx-js/loader']
       }
-    ]
-  }
+    ],
+  },
 };
