@@ -8,9 +8,9 @@ describe('Interface internationalization', () => {
 
   before(() => {
     cy.fixture('users.json').then((users) => {
-      const user = users[0]
+      const user = users.moderator
       cy.createConvo(user.email, user.password)
-      cy.url('pathname').then((adminPath) => {
+      cy.location('pathname').then((adminPath) => {
         const convoPath = adminPath.replace('/m/', '/')
         cy.wrap(convoPath).as('convoPath')
       })
