@@ -25,23 +25,30 @@
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
 
 Cypress.Commands.add("logout", () => {
+  cy.server()
+  cy.route('POST', Cypress.config().apiPath + '/auth/deregister')
+    .as('authLogout')
+
   cy.visit('/signout')
+
+  cy.wait('@authLogout').its('status').should('eq', 200)
 })
 
 Cypress.Commands.add("signup", (name, email, password) => {
   cy.visit('/createuser')
 
   cy.get('form').within(function () {
-    cy.get('input[placeholder=name]').type(name)
-    cy.get('input[placeholder=email]').type(email)
-    cy.get('input[placeholder=password]').type(password)
-    cy.get('input[placeholder="repeat password"]').type(password)
+    cy.get('input#createUserNameInput').type(name)
+    cy.get('input#createUserEmailInput').type(email)
+    cy.get('input#createUserPasswordInput').type(password)
+    cy.get('input#createUserPasswordRepeatInput').type(password)
 
-    cy.get('button').click()
+    cy.get('button#createUserButton').click()
   })
 })
 
 Cypress.Commands.add("login", (email, password) => {
+  cy.logout()
   cy.visit('/signin')
 
   cy.server()
@@ -51,10 +58,10 @@ Cypress.Commands.add("login", (email, password) => {
   }).as('authLogin')
 
   cy.get('form').within(function () {
-    cy.get('input[placeholder=email]').type(email)
-    cy.get('input[placeholder=password]').type(password)
+    cy.get('input#signinEmailInput').type(email)
+    cy.get('input#signinPasswordInput').type(password)
 
-    cy.get('button').click()
+    cy.get('button#signinButton').click()
   })
 
   cy.wait('@authLogin').its('status').should('eq', 200)
@@ -64,5 +71,13 @@ Cypress.Commands.add("createConvo", (adminEmail, adminPassword) => {
   cy.login(adminEmail, adminPassword)
   cy.visit('/')
 
-  cy.get('#root').get('span').contains('New').click()
+  cy.server()
+  cy.route('GET', Cypress.config().apiPath + '/conversations**')
+    .as('getNewConvo')
+
+  cy.get('button').contains('Create new conversation').click()
+
+  cy.wait('@getNewConvo').its('status').should('eq', 200)
+  // Wait for header of convo admin page to be available.
+  cy.contains('h3', 'Configure')
 })
